GC_analysis/requirements.txt
Biopython
pyBigWig
setuptools
